Se refactoriza codigo y además se añade readme

+++ src/main/java/pe/scotiabank/ms/alumnos/service/infrastructure/repository/SpringDataAlumnoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import pe.scotiabank.ms.alumnos.service.infrastructure.repository.entity.AlumnoEntity;
 import reactor.core.publisher.Flux;
 
-public interface SpringDataAlumnoRepository  extends ReactiveCrudRepository<AlumnoEntity, Integer> {
+public interface SpringDataAlumnoRepository  extends ReactiveCrudRepository<AlumnoEntity, Long> {
 
     Flux<AlumnoEntity> findByEstado(Integer estado);
